+++ packages/react/src/components/Authenticator/withAuthenticator.tsx
@@ -1,18 +1,19 @@
 import React from 'react';
-import { Authenticator, AuthenticatorProps } from './Authenticator';
-import { UseAuthenticator } from './hooks/useAuthenticator';
+import { Authenticator, AuthenticatorProps, SignOut } from './Authenticator';
+import { AmplifyUser } from '@aws-amplify/ui';
 
 export type WithAuthenticatorOptions = Omit<AuthenticatorProps, 'children'>;
 
-type AuthenticatorHOCProps = Partial<
-  Pick<UseAuthenticator, 'signOut' | 'user'>
->;
+export interface WithAuthenticatorProps {
+  signOut?: SignOut;
+  user?: AmplifyUser;
+}
 
 /**
  * [📖 Docs](https://ui.docs.amplify.aws/react/connected-components/authenticator)
  */
 export function withAuthenticator<Props = {}>(
-  Component: React.ComponentType<Props & AuthenticatorHOCProps>,
+  Component: React.ComponentType<Props & WithAuthenticatorProps>,
   options: WithAuthenticatorOptions = {}
 ): (props: Props) => JSX.Element {
   const { variation = 'modal' } = options;
@@ -20,8 +21,8 @@ export function withAuthenticator<Props = {}>(
   return function WrappedWithAuthenticator(props: Props) {
     return (
       <Authenticator variation={variation} {...options}>
-        {(authenticatorHOCProps: AuthenticatorHOCProps) => (
-          <Component {...props} {...authenticatorHOCProps} />
+        {(withAuthenticatorProps: WithAuthenticatorProps) => (
+          <Component {...props} {...withAuthenticatorProps} />
         )}
       </Authenticator>
     );