+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/ui-react
 
+## 3.5.4
+
+### Patch Changes
+
+- [#2623](https://github.com/aws-amplify/amplify-ui/pull/2623) [`8afed9306`](https://github.com/aws-amplify/amplify-ui/commit/8afed9306e2e7acc6ce42b20b9a9af2dd70f0ef1) Thanks [@calebpollman](https://github.com/calebpollman)! - fix(ui-react): fix withAuthenticator signOut function typing
+
+## 3.5.3
+
+### Patch Changes
+
+- [#2616](https://github.com/aws-amplify/amplify-ui/pull/2616) [`83bcc0844`](https://github.com/aws-amplify/amplify-ui/commit/83bcc0844eb1049ab49ff4f79280605ef31230d6) Thanks [@calebpollman](https://github.com/calebpollman)! - chore(ui-react, ui): expose Authenticator types, add AmplifyUser interface
+
+- Updated dependencies [[`83bcc0844`](https://github.com/aws-amplify/amplify-ui/commit/83bcc0844eb1049ab49ff4f79280605ef31230d6)]:
+  - @aws-amplify/ui@3.13.4
+
 ## 3.5.2
 
 ### Patch Changes